# data/processed/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ErgoStudy AI Local API",
    "summary": "Deterministic one-day study planning and grounded explanations.",
    "description": "The deterministic planner owns all schedule values. Explanation endpoints may use local Ollama and always retain a validated deterministic fallback.",
    "version": "1.0.0-prototype"
  },
  "paths": {
...
    "/api/v1/plans": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Create one deterministic daily plan",
        "description": "Fast deterministic endpoint. It never calls Ollama.",
        "operationId": "create_plan_api_v1_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Deterministic services unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/full": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Create a deterministic plan",
        "description": "Fast deterministic convenience endpoint. It never calls Ollama.",
        "operationId": "full_plan_api_v1_plans_full_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FullPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FullPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Deterministic services unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explanations": {
      "post": {
        "tags": [
          "explanations"
        ],
        "summary": "Explain an existing deterministic plan",
        "description": "Potentially slow local-LLM endpoint. Timeout, connection, malformed-output, and validation failures return a deterministic fallback with HTTP 200.",
        "operationId": "explanation_api_v1_explanations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplanationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplanationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Deterministic services unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/full-with-explanation": {
      "post": {
        "tags": [
          "demos"
        ],
        "summary": "Run the full demo pipeline",
        "description": "Demo-only and slower than /api/v1/plans/full because it may wait for local Ollama. The plan remains valid when explanation generation falls back.",
        "operationId": "full_with_explanation_api_v1_plans_full_with_explanation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FullWithExplanationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FullWithExplanationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Deterministic services unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DailyStudyPlanSchema": {
        "properties": {
          "plan_id": {
            "type": "string",
            "maxLength": 240,
            "minLength": 1,
            "title": "Plan Id"
          },
          "total_available_minutes": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Total Available Minutes"
          },
          "total_study_minutes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Study Minutes"
          },
          "total_break_minutes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Break Minutes"
          },
          "total_planned_minutes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Planned Minutes"
          },
          "unallocated_minutes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unallocated Minutes"
          },
          "scheduled_subjects": {
            "items": {
              "$ref": "#/components/schemas/ScheduledSubjectSchema"
            },
            "type": "array",
            "title": "Scheduled Subjects"
          },
          "unscheduled_subjects": {
            "items": {
              "$ref": "#/components/schemas/UnscheduledSubjectSchema"
            },
            "type": "array",
            "title": "Unscheduled Subjects"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/PlannedSessionSchema"
            },
            "type": "array",
            "title": "Sessions"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "planner_version": {
            "type": "string",
            "maxLength": 240,
            "minLength": 1,
            "title": "Planner Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan_id",
          "total_available_minutes",
          "total_study_minutes",
          "total_break_minutes",
          "total_planned_minutes",
          "unallocated_minutes",
          "scheduled_subjects",
          "unscheduled_subjects",
          "sessions",
          "warnings",
          "planner_version"
        ],
        "title": "DailyStudyPlanSchema"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error",
          "request_id"
        ],
        "title": "ErrorResponse"
      },
      "ExplanationRequest": {
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/DailyStudyPlanSchema"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan"
        ],
        "title": "ExplanationRequest"
      },
      "ExplanationResponse": {
        "properties": {
          "grounded_response": {
            "$ref": "#/components/schemas/GroundedResponseSchema"
          },
          "generation_mode": {
            "type": "string",
            "enum": [
              "local_llm",
              "local_llm_corrected",
              "deterministic_fallback"
            ],
            "title": "Generation Mode"
          },
          "model_status": {
            "type": "string",
            "enum": [
              "available",
              "unavailable",
              "timed_out",
              "invalid_output",
              "failed"
            ],
            "title": "Model Status"
          },
          "validation_status": {
            "type": "string",
            "enum": [
              "valid",
              "corrected_valid",
              "fallback_valid"
            ],
            "title": "Validation Status"
          },
          "fallback_reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason Code"
          },
          "attempt_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Attempt Count"
          },
          "generation_latency_ms": {
            "type": "number",
            "title": "Generation Latency Ms"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "v1"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "grounded_response",
          "generation_mode",
          "model_status",
          "validation_status",
          "fallback_reason_code",
          "attempt_count",
          "generation_latency_ms",
          "request_id"
        ],
        "title": "ExplanationResponse"
      },
      "FallbackInformation": {
        "properties": {
          "used": {
            "type": "boolean",
            "title": "Used"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reason Codes"
          },
          "subjects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subjects"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "used"
        ],
        "title": "FallbackInformation"
      },
      "FullPlanRequest": {
        "properties": {
          "total_available_minutes": {
            "type": "integer",
            "maximum": 720.0,
            "minimum": 30.0,
            "title": "Total Available Minutes"
          },
          "preferred_start_time": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Start Time"
          },
          "preferred_session_length": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 60.0,
                "minimum": 20.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Session Length"
          },
          "subjects": {
            "items": {
              "$ref": "#/components/schemas/SubjectRequest"
            },
            "type": "array",
            "maxItems": 30,
            "minItems": 1,
            "title": "Subjects"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "total_available_minutes",
          "subjects"
        ],
        "title": "FullPlanRequest"
      },
      "FullPlanResponse": {
        "properties": {
          "original_plan": {
            "$ref": "#/components/schemas/DailyStudyPlanSchema"
          },
          "final_sessions": {
            "items": {
              "$ref": "#/components/schemas/PlannedSessionSchema"
            },
            "type": "array",
            "title": "Final Sessions"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "processing_time_ms": {
            "type": "number",
            "title": "Processing Time Ms"
          },
          "fallback": {
            "$ref": "#/components/schemas/FallbackInformation"
          },
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "v1"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "original_plan",
          "final_sessions",
          "request_id",
          "processing_time_ms",
          "fallback"
        ],
        "title": "FullPlanResponse"
      },
      "FullWithExplanationRequest": {
        "properties": {
          "total_available_minutes": {
            "type": "integer",
            "maximum": 720.0,
            "minimum": 30.0,
            "title": "Total Available Minutes"
          },
          "preferred_start_time": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Start Time"
          },
          "preferred_session_length": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 60.0,
                "minimum": 20.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Session Length"
          },
          "subjects": {
            "items": {
              "$ref": "#/components/schemas/SubjectRequest"
            },
            "type": "array",
            "maxItems": 30,
            "minItems": 1,
            "title": "Subjects"
          },
          "explanation_timeout_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation Timeout Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "total_available_minutes",
          "subjects"
        ],
        "title": "FullWithExplanationRequest"
      },
      "FullWithExplanationResponse": {
        "properties": {
          "original_plan": {
            "$ref": "#/components/schemas/DailyStudyPlanSchema"
          },
          "explanation": {
            "$ref": "#/components/schemas/ExplanationResponse"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "processing_time_ms": {
            "type": "number",
            "title": "Processing Time Ms"
          },
          "fallback": {
            "$ref": "#/components/schemas/FallbackInformation"
          },
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "v1"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "original_plan",
          "explanation",
          "request_id",
          "processing_time_ms",
          "fallback"
        ],
        "title": "FullWithExplanationResponse"
      },
      "GroundedResponseSchema": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "allocation_explanations": {
            "items": {
              "$ref": "#/components/schemas/AllocationExplanationSchema"
            },
            "type": "array",
            "title": "Allocation Explanations"
          },
          "session_messages": {
            "items": {
              "$ref": "#/components/schemas/SessionMessageSchema"
            },
            "type": "array",
            "title": "Session Messages"
          },
          "unscheduled_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unscheduled Message"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary",
          "allocation_explanations",
          "session_messages",
          "unscheduled_message",
          "warnings"
        ],
        "title": "GroundedResponseSchema"
      },
...
      "PlanRequest": {
        "properties": {
          "total_available_minutes": {
            "type": "integer",
            "maximum": 720.0,
            "minimum": 30.0,
            "title": "Total Available Minutes"
          },
          "preferred_start_time": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Start Time"
          },
          "preferred_session_length": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 60.0,
                "minimum": 20.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Session Length"
          },
          "subjects": {
            "items": {
              "$ref": "#/components/schemas/SubjectRequest"
            },
            "type": "array",
            "maxItems": 30,
            "minItems": 1,
            "title": "Subjects"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "total_available_minutes",
          "subjects"
        ],
        "title": "PlanRequest"
      },
      "PlanResponse": {
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/DailyStudyPlanSchema"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "processing_time_ms": {
            "type": "number",
            "title": "Processing Time Ms"
          },
          "fallback": {
            "$ref": "#/components/schemas/FallbackInformation"
          },
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "v1"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan",
          "request_id",
          "processing_time_ms",
          "fallback"
        ],
        "title": "PlanResponse"
      },
      "PlannedSessionSchema": {
        "properties": {
          "order": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Order"
          },
          "session_type": {
            "type": "string",
            "enum": [
              "study",
              "break"
            ],
            "title": "Session Type"
          },
          "duration_minutes": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Duration Minutes"
          },
          "reason": {
            "type": "string",
            "maxLength": 240,
            "minLength": 1,
            "title": "Reason"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "cognitive_demand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cognitive Demand"
          },
          "recommended_methods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Methods"
          },
          "retrieved_record_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Retrieved Record Ids"
          },
          "used_fallback": {
            "type": "boolean",
            "title": "Used Fallback",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "order",
          "session_type",
          "duration_minutes",
          "reason"
        ],
        "title": "PlannedSessionSchema"
      },
...
      "SubjectRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "topics": {
            "items": {
              "type": "string",
              "maxLength": 160,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 20,
            "title": "Topics"
          },
          "difficulty": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Difficulty"
          },
          "priority": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Priority"
          },
          "workload": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Workload"
          },
          "current_understanding": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Current Understanding"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "difficulty",
          "priority",
          "workload",
          "current_understanding"
        ],
        "title": "SubjectRequest"
      },
...
    }
  }
}